+++ CHANGELOG.md
@@ -12,10 +12,12 @@ New release of Oragono!
 
 ### Added
 * Added `REHASH` command.
+* Added ability to message channel members with a specific privelege (i.e. support for `STATUSMSG`).
 * Added ability to enable and disable SASL.
 * Added support for IRCv3 capabilities [`cap-notify`](http://ircv3.net/specs/extensions/cap-notify-3.2.html) and [`echo-message`](http://ircv3.net/specs/extensions/echo-message-3.2.html).
 
 ### Changed
+* Server operators no longer have permissions to do everything in channels. 
 
 ### Removed